+++ packages/package-json-lint-config-terra/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 1.2.0 - (October 12, 2021)
+
+* Added
+  * Added new rule `require-ie10-compatible-dependency-versions`.
+  * Added new rule `require-dependencies-declared-at-appropriate-level`.
+
 ## 1.1.0 - (September 28, 2021)
 
 * Changed