+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+## 2.6.0 - (October 12, 2021)
+
+* Added
+  * Added documentation for rule `require-ie10-compatible-dependency-versions`.
+  * Added documentation for new rule `require-dependencies-declared-at-appropriate-level`.
+
 ## 2.5.2 - (September 30, 2021)
 
 * Added